feat: modularize src/index.ts into separate modules
Refactored 693-line index.ts into focused modules for better code organization and maintainability:
- types.ts: Shared TypeScript interfaces
- config.ts: Configuration and validation logic
- client.ts: CodecovClient API class
- tools.ts: Tool definitions and handlers
- prompts.ts: Prompt definitions and handlers
- resources.ts: Resource definitions and content
- server.ts: MCP server initialization
- index.ts: Entry point with re-exports (57 lines)

Benefits:
- Clear separation of concerns
- Easier navigation and maintenance
- Better code organization for publication
- Maintained backward compatibility
- All 69 tests passing with 95.49% coverage

Resolves #60
